survivors#notify_infection API
This commit includes survivors#notify_infection API.
It was added the InfectionNotification model to register
the notifier and the notified.

If there is three notifications to one infected survivor
it will be flagged as infected.

It was added the gem Interactor (https://github.com/collectiveidea/interactor)
its propose is to encapsulate the application business logic
which is usually put in the controller making it fat.